Redo handling of vars in parser
Shift handling of module qualifications on vars from parser into lexer.
Working out if a name is qualified is now a job for the lexer. Reduced
the mess of different sorts of vars to just 4 qCon pCon qVar pVar, for 
qualified and unqualified Cons or var/symbol. 

Should probably reduce this to have just 'con' and 'var' and let the
renamer check whether it's valid to have a qualifier in a particular
place - that way we don't have to stop the parser straight away on error.