Refactor representation of type expressions
We now always use a general type application form TApp 
instead of constructors specific to functions (TFun) and
data types (TData). This makes the inferencer somewhat
slower as there are more nodes in the graph, but also makes
it simpler, more sane, and easier to work on.

Inferencer runtime performance to be addressed in a
subsequent patch.