Fix #99 : Finding which modules need to be rebuilt is inefficient.
Profing of the build of the Rover program in the test suite show that as much
as 90% of the run time and 90% of the memory usage was used just to figure
out which modules needed to be rebuilt.

The problem turned out to be an O(n^2) (and possibly worse) algorithm which
tried to build the ScrapeGraph and figure out which modules needed to be
rebuild in a single pass. Separating this into two passed, one to build the
ScrapeGraph and one to propagate the NeedsRebuild flag fixed the problem.
Building the ScrapeGraph now takes less than 1% of the compile run time.