Allow module ids at the start of modules
I decided to use:

  module Graphics.Raster.Triangle
  import Data.List
  thing1 = ...
  thing2 = ...
  
instead of the Haskell standard:

  module Graphics.Raster.Triangle where
  import Data.List
  thing1 = ...
  thing2 = ...
  
That 'where' is funny because the bindings do not start to the right
of 'module', which makes it unlike any other 'where' in the program. 
  
If we were going to be consistent we'd have

  module Graphics.Raster.Triangle where
      import Data.List
      thing1 = ...
      thing2 = ...
  
With the bindings indented, but we don't want the whole program
indented.
  
If we want to support nested modules in the same source file at
a later date then this last version would be easy to add.
  
I hacked around for a while trying to get the Haskell standard 
version to work, but I'd have to handle it as a specicial case
and rewrite other code. Taking that as an omen I gave up on it.
  
We don't actually do anything with the module ids yet, 
besides parse them. Patch pending.