During data elaboration, link in effects of funs in constructors
For a data type like
 data Foo a b
       = MkFoo (a -> b)

This used to force the embedded function to be pure, but I think
it's better if it just elaborates to:

 data Foo a b !e1 $c1
       = MkFoo (a -(!e1 $c1)> b)

This way there should be less of a chance of having to worry about the 
effect or closure annots. We could still require the contained function
to be pure by adding a constraint directly to the alternative.