Add nifty higher kinded peekOn and pokeOn fns
 pokeOn	:: forall (t :: % -> *) a
	.  t %r1 -> Ptr# a -> a -(!e1)> ()
	:- !e1 = !Write %r1

 peekOn	:: forall (t :: % -> *) a
	.  t %r1 -> Ptr# a -(!e1)> a
	:- !e1 = !Read %r1

These are used to read and write to a pointer, while assigning
the Read or Write effect to the primary region of some other type.

The fancy part is that by using the higher kinded binding variable
(t :: % -> *) we can ignore what constructor the type of interest
is actually using.