LLVM : Add method for defining structs.
Need a way of defining structs so that the generated LLVM code will be
compatible with the structs in the C runtime. Since LLVM structs are packed
and the C structs are padded, padding needs to be added manually to the
LLVM versions and the padding will need to be different depending on whether
the platform is 32 or 64 bit.

However, since LLVM structs are accessed by index, adding padding (which
ie may be present on 64 bit and absent on 32 bit) messes up the indexing;
the indexing of any given field may be different on 32 and 64 bit.

The solution is defined and AbstractStruct specification that gets converted
to a system dependant opaque LlvmStructDesc data type and adding functions
that allow querying the LlvmStructDesc by field name to get the field index
and type.