Use weak form types in all the type class definitions
"weak form" means the effects of functions are written like this:
 foo :: forall %r1 !e1. Int %r1 -(!e1)> Int %r1
     :- !e1 :> Read %r1

Instead of the "strong form" like this:
 foo :: forall %r1. Int %r1 -(!e1)> Int %r1
     :- !e1 = Read %r1

The polymorphic effect system upon with Disciple's system is based uses
the weak form for inference, but we want to use the strong form in core
because it means there's less effect information to pass around. At
the moment I've been strenghtening types during inference, but it turns
out that's the reason the Order4 (4th order function) tests are failing.

Strengthening types only after inference is complete should fix these 
tests, but there's some rearranging to be done to make the rest of 
the compiler cope with the weak form types. This is the first step.