rewrites: foldr/build wasn't working for lists starting with 'build' Foldr/build would not fuse in a case like
> map double (count 100)
==
> build (/\(e : !)
>	\(k : a -> b -(e)> b)
>	\(z : b).
>	    foldr ... (build ...))

because the inner-most foldr has an uninstantiated effect type of e, so it can't tell whether it's disjoint.

I ended up adding an extra function, foldrAndBuild, and a rule which fuses a foldr/build pair into
a single call to foldrAndBuild, regardless of disjointness.
After that, we can safely inline any remaining instances of 'build' since they would not fuse.
Then simplify the expression and apply a new rule that fuses a foldrAndBuild call only if its effects are disjoint.