ddc-core-flow: to-tetra: a-normalise before converting
a-normalise before converting, because the conversion can actually
duplicate occurrences of case scrutinees, and other arguments.
a-normalisation beforehand will assure that these arguments to
primitives and the scrutinees will just be variables, so no duplication
of work is possible.

this a-normalisation was being done before, but floating was done
afterwards, which could somewhat undo this.
instead, the floating is now deferred until afterwards.