Recompilation and dependencies: sed to the rescue!
Because GHC touches the .hi after the .o, adding a build dependency from
the .hi to the .o can cause spurious recompilations.
It might not happen all the time, or it might happen often, depending on
your clock resolution and time ghc takes between outputting the two
files.

So instead, suggest modifying the dependency files that GHC produces so
that the .o depends on another .o. That way we don't have to worry about
order between them or anything, and it should make your build faster.