salt: purge rotted C code generator.
At one time we converted Salt directly to C, but after implementation of the LLVM
backend this code has rotted. If we want a C code generator back it would be easier
to convert the output of the LLVM generator to C, rather than fixing the old C code
generator.